move jwt to adapter/application package
