+++ src/main/java/com/koc/user/application/service/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.koc.user.application.jwt;
+package com.koc.user.application.service;
 
 
 import io.jsonwebtoken.Claims;
