+++ src/main/java/com/koc/user/domain/UserToken.java
@@ -1,6 +1,6 @@
 package com.koc.user.domain;
 
-import com.koc.user.application.jwt.JwtProvider;
+import com.koc.user.application.service.JwtProvider;
 import com.koc.user.adapter.out.persistence.UserTokenEntity;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
